% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids.R
\name{get_ids}
\alias{get_ids}
\title{Legislators IDs data.}
\format{Data frame with columns (might vary by legislature):
\itemize{
\item{wikidataid: Wikidata ID identifying a legislator's Wikidata entry (of class \sQuote{character}).}
\item{parlid: Parliamentary website ID  (of class \sQuote{character}).}
\item{sycomore: Sycomore database of French MPs (of class \sQuote{character}).}
\item{libcon: Library of Congress ID (of class \sQuote{character}).}
\item{gnd: German National Library ID (of class \sQuote{character}).}
\item{bnf: French National Library ID (of class \sQuote{character}).}
\item{freebase: Freebase ID (of class \sQuote{character}).}
\item{munzinger:  Munzinger archive ID (of class \sQuote{character}).}
\item{nndb: Notable Names Database ID (of class \sQuote{character}).}
\item{imdb: Internet Movie Database ID (of class \sQuote{character}).}
\item{brittanica: Encyclopedia Brittanica ID (of class \sQuote{character}).}
\item{quora: Quora ID (of class \sQuote{character}).}
\item{votesmart: Project Votesmart ID (of class \sQuote{character}).}
\item{fec: Federal Election Commission ID (of class \sQuote{character}).}
\item{ballotpedia: Ballotpedia ID (of class \sQuote{character}).}
\item{opensecrets: Opensecrets ID (of class \sQuote{character}).}
\item{genealogists: Genealogists ID (of class \sQuote{character}).}
\item{politfacts: Politfacts ID (of class \sQuote{character}).}
}}
\source{
Wikipedia API, \url{https://wikipedia.org/w/api.php}
}
\usage{
get_ids(legislature)
}
\arguments{
\item{legislature}{A character string specifying the legislature for which data shall be fetched from the GitHub repository. Currently one of \sQuote{austria}, \sQuote{france}, \sQuote{germany}, \sQuote{ireland}, \sQuote{usah}, or \sQuote{usas}.}
}
\value{
A data frame with columns as specified above.
}
\description{
Fetches a range of IDs of legislators for the specified legislature from the GitHub repository. Requires a working Internet connection.
}
\examples{
\donttest{## assign entire core dataset into the environment
at_elites <- get_core(legislature = "austria")

## assign only data for the 12th legislative session into the environment
at_elites_subset <- dplyr::semi_join(x = get_core(legislature = "austria"),
                                     y = dplyr::filter(get_political(legislature = "austria"),
                                                                     session == 8),
                                     by = "pageid")

## join at_elites_subset with respective history dataset
at_history <- dplyr::left_join(x = at_elites_subset,
                               y = get_history(legislature = "austria"),
                               by = "pageid")

## assign only birthdate for members of the political party 'SdP' into the environment
at_birthdates_SdP <- dplyr::semi_join(x = dplyr::select(get_core(legislature = "austria"),
                                                        pageid, birth),
                                      y = dplyr::filter(get_political(legislature = "austria"),
                                                                      party == "SdP"),
                                      by = "pageid")$birth
}
}
