% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_history.R
\name{get_history}
\alias{get_history}
\title{Fetch Wikipedia 'History' dataset}
\format{Data frame with columns (might vary by legislature):
\itemize{
\item{pageid: Wikipedia page ID identifying a legislator's Wikipedia biography (of class \sQuote{integer}).}
\item{revid: Wikipedia edit ID (of class \sQuote{integer}).}
\item{parentid: Wikipedia edit ID of the previous revision (of class \sQuote{integer}).}
\item{user: Username of registered user responsible for the revision, IP address in case of anonymous revision (of class \sQuote{character}).}
\item{userid: ID of registered user responsible for the revision, 0 in case of anonymous revision (of class \sQuote{integer}).}
\item{timestamp: Date and time of the revision (of class \sQuote{POSIXct}).}
\item{size: Revision size in bytes (of class \sQuote{integer}).}
\item{comment: Revision comment (of class \sQuote{character}).}
}}
\source{
Wikipedia API, \url{https://wikipedia.org/w/api.php}
}
\usage{
get_history(legislature)
}
\arguments{
\item{legislature}{A character string specifying the legislature for which data shall be fetched. Currently one of \sQuote{aut}, \sQuote{can}, \sQuote{cze}, \sQuote{fra}, \sQuote{deu}, \sQuote{irl}, \sQuote{sco}, \sQuote{gbr}, \sQuote{usa_house}, or \sQuote{usa_senate}.}
}
\value{
A data frame with columns as specified above.
}
\description{
Fetches full revision histories of legislators' Wikipedia biographies for the specified legislature. Requires a working Internet connection.
}
\examples{
\donttest{## assign entire core dataset into the environment
aut_politicians <- get_core(legislature = "aut")

## assign only data for the 12th legislative session into the environment
aut_politicians_subset <- dplyr::semi_join(x = get_core(legislature = "aut"),
                                           y = dplyr::filter(get_political(legislature = "aut"),
                                                                     session == 8),
                                           by = "pageid")

## join aut_politicians_subset with respective history dataset
aut_politicians_history <- dplyr::left_join(x = aut_politicians_subset,
                                            y = get_history(legislature = "aut"),
                                            by = "pageid")

## assign only birthdate for members of the political party 'SdP' into the environment
aut_birthdates_SdP <- dplyr::semi_join(x = dplyr::select(get_core(legislature = "aut"),
                                                         pageid, birth),
                                       y = dplyr::filter(get_political(legislature = "aut"),
                                                                       party == "SdP"),
                                       by = "pageid")$birth
}
}
