% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_political.R
\name{get_political}
\alias{get_political}
\title{Legislators political data.}
\format{Data frame in long format with columns (might vary by legislature):
\itemize{
\item{pageid: Wikipedia page ID identifying a legislator's Wikipedia biography (of class \sQuote{integer}).}
\item{session: Legislative period (of class \sQuote{character}).}
\item{party: A legislator's party affiliation (of class \sQuote{character}).}
\item{constituency: A legislator's constituency (of class \sQuote{character}).}
\item{session_start: Date the legislative period started (of class \sQuote{Date}).}
\item{session_end: Date the legislative period ended (of class \sQuote{Date}).}
\item{service: A legislator's period of service in days during the respective session (of class \sQuote{difftime})}
}}
\source{
Wikipedia, \url{https://wikipedia.org/} \cr
}
\usage{
get_political(legislature)
}
\arguments{
\item{legislature}{A character string specifying the legislature for which data shall be fetched from the GitHub repository. Currently one of \sQuote{austria}, \sQuote{france}, \sQuote{germany}, \sQuote{ireland}, \sQuote{usah}, or \sQuote{usas}.}
}
\value{
A data frame with columns as specified above.
}
\description{
Fetches political data for legislators of the specified legislature from the GitHub repository. Requires a working Internet connection.
}
\examples{
\donttest{## assign entire core dataset into the environment
at_elites <- get_core(legislature = "austria")

## assign only data for the 12th legislative session into the environment
at_elites_subset <- dplyr::semi_join(x = get_core(legislature = "austria"),
                                     y = dplyr::filter(get_political(legislature = "austria"),
                                                                     session == 8),
                                     by = "pageid")

## join at_elites_subset with respective history dataset
at_history <- dplyr::left_join(x = at_elites_subset,
                               y = get_history(legislature = "austria"),
                               by = "pageid")

## assign only birthdate for members of the political party 'SdP' into the environment
at_birthdates_SdP <- dplyr::semi_join(x = dplyr::select(get_core(legislature = "austria"),
                                                        pageid, birth),
                                      y = dplyr::filter(get_political(legislature = "austria"),
                                                                      party == "SdP"),
                                      by = "pageid")$birth
}
}
