% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_political.R
\name{get_political}
\alias{get_political}
\title{Fetch 'Political' table}
\format{Data frame in long format with columns (varies by legislature):
\itemize{
\item{pageid: Wikipedia page ID identifying a legislator's Wikipedia biography (of class \sQuote{integer}).}
\item{session: Legislative period (of class \sQuote{integer}).}
\item{party: A legislator's party affiliation (of class \sQuote{character}). See \url{https://github.com/saschagobel/legislatoR} for the full form of abbreviated party names and english translations of non-english party names}
\item{constituency: A legislator's constituency (of class \sQuote{character}).}
\item{constituency2: A legislator's constituency (upper level, if applicable, of class \sQuote{character}).}
\item{constituency_id: ID of a legislator's constituency (of class \sQuote{character}).}
\item{session_start: Date the legislative period started (of class \sQuote{Date}).}
\item{session_end: Date the legislative period ended (of class \sQuote{Date}).}
\item{service: A legislator's period of service in days during the respective session (of class \sQuote{integer})}
\item{government (or similar): Indicator of a legislator's majority status in parliament (of class \sQuote{logical}). Further columns with extensions of this might exist.}
\item{leader (or similar): Indicator of a legislator's leader status in parliament (of class \sQuote{logical}). Further columns with extensions of this might exist.}
}}
\source{
Wikipedia, \url{https://wikipedia.org/} \cr
Czech Republic Parliamentary Members Archive \url{http://public.psp.cz/sqw/fsnem.sqw?zvo=1} \cr
Spain Parliamentary Members Archive \url{http://www.congreso.es/portal/page/portal/Congreso/Congreso/Diputados}
}
\usage{
get_political(legislature)
}
\arguments{
\item{legislature}{A character string specifying the three-letter country code of the legislature for which data shall be fetched. Currently one of \sQuote{aut}, \sQuote{can}, \sQuote{cze}, \sQuote{esp}, \sQuote{fra}, \sQuote{deu}, \sQuote{irl}, \sQuote{sco}, \sQuote{gbr}, \sQuote{usa_house}, or \sQuote{usa_senate}.}
}
\value{
A data frame with columns as specified above.
}
\description{
Fetches political data of legislators for the specified legislature. Requires a working Internet connection.
}
\examples{
\donttest{# Get entire 'Political' table for the Czech Poslanecka Snemovna
cze_political <- get_political(legislature = "cze")
tibble::glimpse(cze_political)

# Get 'Political' table for female DSP party members of the Czech Poslanecka Snemovna
cze_political_subset <- dplyr::semi_join(x = dplyr::filter(cze_political,
                                                           party == "ODS"),
                                         y = dplyr::filter(get_core(legislature = "cze"),
                                                           sex == "female"),
                                         by = "pageid")
tibble::glimpse(cze_political_subset)
}
}
