% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_social.R
\name{get_social}
\alias{get_social}
\title{Fetch 'Social' dataset}
\format{Data frame with columns (might vary by legislature):
\itemize{
\item{wikidataid: Wikidata ID identifying a legislator's Wikidata entry (of class \sQuote{character}).}
\item{twitter: Twitter handle (of class \sQuote{character}).}
\item{facebook: Facebook handle (of class \sQuote{character}).}
\item{youtube: Youtube ID (of class \sQuote{character}).}
\item{googlep: Google Plus ID (of class \sQuote{character}).}
\item{instagram: Instagram handle (of class \sQuote{character}).}
\item{linkedin: LinkedIn ID (of class \sQuote{character}).}
\item{website: Personal website URL (of class \sQuote{character}).}
}}
\source{
Wikidata API, \url{https://www.wikidata.org/}
}
\usage{
get_social(legislature)
}
\arguments{
\item{legislature}{A character string specifying the legislature for which data shall be fetched. Currently one of \sQuote{aut}, \sQuote{can}, \sQuote{cze}, \sQuote{fra}, \sQuote{deu}, \sQuote{irl}, \sQuote{sco}, \sQuote{gbr}, \sQuote{usa_house}, or \sQuote{usa_senate}.}
}
\value{
A data frame with columns as specified above.
}
\description{
Fetches social media handles and website URLs of legislators for the specified legislature. Requires a working Internet connection.
}
\examples{
\donttest{## assign entire core dataset into the environment
aut_politicians <- get_core(legislature = "aut")

## assign only data for the 12th legislative session into the environment
aut_politicians_subset <- dplyr::semi_join(x = get_core(legislature = "aut"),
                                           y = dplyr::filter(get_political(legislature = "aut"),
                                                                     session == 8),
                                           by = "pageid")

## join aut_politicians_subset with respective history dataset
aut_politicians_history <- dplyr::left_join(x = aut_politicians_subset,
                                            y = get_history(legislature = "aut"),
                                            by = "pageid")

## assign only birthdate for members of the political party 'SdP' into the environment
aut_birthdates_SdP <- dplyr::semi_join(x = dplyr::select(get_core(legislature = "aut"),
                                                         pageid, birth),
                                       y = dplyr::filter(get_political(legislature = "aut"),
                                                                       party == "SdP"),
                                       by = "pageid")$birth
}
}
